1、完善内容显示 2、增加异常判断 3、蜂鸣器控制和屏幕亮度控制

+++ src/App/W25Q64.h
@@ -12,7 +12,7 @@ void W25Q64PowerDown(void);
 void W25Q64Wakeup(void);
 
 uint16_t W25Q64ReadID(void);
-int W25Q64Initialize(void);
+uint16_t W25Q64Initialize(void);
 void W25Q64Poll(void);
 
 #endif
